Rename submodules/replicode to AERA. See pull request #8.
