+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "submodules/replicode"]
-	path = submodules/replicode
+[submodule "submodules/AERA"]
+	path = submodules/AERA
 	url = https://github.com/IIIM-IS/AERA
