+++ graphics-items/aera-visualizer-scene.cpp
@@ -50,7 +50,7 @@
 //_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/
 
 #include <algorithm>
-#include "../submodules/replicode/r_exec/factory.h"
+#include "../submodules/AERA/r_exec/factory.h"
 #include "aera-visualizer-window.hpp"
 #include "arrow.hpp"
 #include "anchored-horizontal-line.hpp"
