+++ graphics-items/composite-state-prediction-item.cpp
@@ -54,7 +54,7 @@
 #include <QRegularExpression>
 #include "../explanation-log-window.hpp"
 #include "../aera-visualizer-window.hpp"
-#include "../submodules/replicode/r_exec/factory.h"
+#include "../submodules/AERA/r_exec/factory.h"
 #include "aera-visualizer-scene.hpp"
 #include "instantiated-composite-state-item.hpp"
 #include "composite-state-prediction-item.hpp"
